Changed rotation data type.

+++ src/main/java/org/evilco/mc/defense/client/model/sensor/CameraModel.java
@@ -70,13 +70,13 @@ public class CameraModel extends ModelBase {
 	 * Renders the camera.
 	 * @see net.minecraft.client.model.ModelBase.render ()
 	 */
-	public void render (Entity entity, float f, float f1, float f2, float f3, float f4, float f5, float cameraAngle) {
+	public void render (Entity entity, float f, float f1, float f2, float f3, float f4, float f5, double cameraAngle) {
 		super.render (entity, f, f1, f2, f3, f4, f5);
 		this.setRotationAngles (f, f1, f2, f3, f4, f5, entity);
 
 		// apply angle
-		this.base.rotateAngleY = cameraAngle;
-		this.lens.rotateAngleY = cameraAngle;
+		this.base.rotateAngleY = ((float) cameraAngle);
+		this.lens.rotateAngleY = ((float) cameraAngle);
 
 		this.base.render (f5);
 		this.lens.render (f5);
